codegen: route payable send/transfer through TargetCodegen::lower_builtin
