+++ src/codegen/targets/mod.rs
@@ -2,9 +2,10 @@
 
 pub(crate) mod soroban;
 
-use crate::codegen::cfg::ControlFlowGraph;
+use crate::codegen::cfg::{ControlFlowGraph, Instr};
 use crate::codegen::events::polkadot::PolkadotEventEmitter;
 use crate::codegen::events::solana::SolanaEventEmitter;
+use crate::codegen::expression::expression;
 use crate::codegen::interface::{EventEmitter, TargetCodegen};
 use crate::codegen::solana_accounts::account_collection::collect_accounts_from_contract;
 use crate::codegen::solana_accounts::account_management::manage_contract_accounts;
@@ -12,9 +13,11 @@ use crate::codegen::storage::{
     array_pop, array_push, storage_slots_array_pop, storage_slots_array_push,
 };
 use crate::codegen::vartable::Vartable;
-use crate::codegen::{dispatch, Expression, Options};
+use crate::codegen::{dispatch, polkadot, Expression, Options};
 use crate::sema::ast;
-use crate::sema::ast::{Function, Namespace, RetrieveType, StructType, Type};
+use crate::sema::ast::{
+    CallTy, ExternalCallAccounts, Function, Namespace, RetrieveType, StructType, Type,
+};
 use crate::Target;
 use num_bigint::BigInt;
 use num_traits::Zero;
@@ -318,6 +321,12 @@ impl TargetCodegen for PolkadotTarget {
                     self,
                 ))
             }
+            ast::Builtin::PayableSend => {
+                Some(self.payable_send(loc, args, cfg, contract_no, func, ns, vartab, opt))
+            }
+            ast::Builtin::PayableTransfer => {
+                Some(self.payable_transfer(loc, args, cfg, contract_no, func, ns, vartab, opt))
+            }
             _ => None,
         }
     }
@@ -349,5 +358,144 @@ impl TargetCodegen for PolkadotTarget {
                 value: offset,
             }),
         }
+    }
+}
+
+impl PolkadotTarget {
+    /// Lower `address.send(value)` (`Builtin::PayableSend`). EVM routes through an external
+    /// call with an empty payload; Polkadot emits a `ValueTransfer` and checks the return code.
+    fn payable_send(
+        &self,
+        loc: &Loc,
+        args: &[ast::Expression],
+        cfg: &mut ControlFlowGraph,
+        contract_no: usize,
+        func: Option<&Function>,
+        ns: &Namespace,
+        vartab: &mut Vartable,
+        opt: &Options,
+    ) -> Expression {
+        let address = expression(&args[0], cfg, contract_no, func, ns, vartab, opt, self);
+        let value = expression(&args[1], cfg, contract_no, func, ns, vartab, opt, self);
+        let success = vartab.temp(
+            &pt::Identifier {
+                loc: *loc,
+                name: "success".to_owned(),
+            },
+            &Type::Uint(32),
+        );
+
+        // Ethereum can only transfer via external call
+        if self.is_evm {
+            cfg.add(
+                vartab,
+                Instr::ExternalCall {
+                    loc: *loc,
+                    success: Some(success),
+                    address: Some(address),
+                    accounts: ExternalCallAccounts::AbsentArgument,
+                    seeds: None,
+                    payload: Expression::AllocDynamicBytes {
+                        loc: *loc,
+                        ty: Type::DynamicBytes,
+                        size: Box::new(Expression::NumberLiteral {
+                            loc: *loc,
+                            ty: Type::Uint(32),
+                            value: BigInt::from(0),
+                        }),
+                        initializer: Some(vec![]),
+                    },
+                    value,
+                    gas: Expression::NumberLiteral {
+                        loc: *loc,
+                        ty: Type::Uint(64),
+                        value: BigInt::from(i64::MAX),
+                    },
+                    callty: CallTy::Regular,
+                    contract_function_no: None,
+                    flags: None,
+                },
+            );
+            return Expression::Variable {
+                loc: *loc,
+                ty: Type::Bool,
+                var_no: success,
+            };
+        }
+
+        cfg.add(
+            vartab,
+            Instr::ValueTransfer {
+                success: Some(success),
+                address,
+                value,
+            },
+        );
+
+        polkadot::check_transfer_ret(loc, success, cfg, ns, opt, vartab, false).unwrap()
+    }
+
+    /// Lower `address.transfer(value)` (`Builtin::PayableTransfer`). EVM routes through an
+    /// external call; Polkadot emits a `ValueTransfer` and reverts on a non-zero return code.
+    fn payable_transfer(
+        &self,
+        loc: &Loc,
+        args: &[ast::Expression],
+        cfg: &mut ControlFlowGraph,
+        contract_no: usize,
+        func: Option<&Function>,
+        ns: &Namespace,
+        vartab: &mut Vartable,
+        opt: &Options,
+    ) -> Expression {
+        let address = expression(&args[0], cfg, contract_no, func, ns, vartab, opt, self);
+        let value = expression(&args[1], cfg, contract_no, func, ns, vartab, opt, self);
+        if self.is_evm {
+            // Ethereum can only transfer via external call
+            cfg.add(
+                vartab,
+                Instr::ExternalCall {
+                    loc: *loc,
+                    success: None,
+                    accounts: ExternalCallAccounts::AbsentArgument,
+                    seeds: None,
+                    address: Some(address),
+                    payload: Expression::AllocDynamicBytes {
+                        loc: *loc,
+                        ty: Type::DynamicBytes,
+                        size: Box::new(Expression::NumberLiteral {
+                            loc: *loc,
+                            ty: Type::Uint(32),
+                            value: BigInt::from(0),
+                        }),
+                        initializer: Some(vec![]),
+                    },
+                    value,
+                    gas: Expression::NumberLiteral {
+                        loc: *loc,
+                        ty: Type::Uint(64),
+                        value: BigInt::from(i64::MAX),
+                    },
+                    callty: CallTy::Regular,
+                    contract_function_no: None,
+                    flags: None,
+                },
+            );
+            return Expression::Poison;
+        }
+
+        let success = vartab.temp_name("success", &Type::Uint(32));
+        cfg.add(
+            vartab,
+            Instr::ValueTransfer {
+                success: Some(success),
+                address,
+                value,
+            },
+        );
+
+        polkadot::check_transfer_ret(loc, success, cfg, ns, opt, vartab, true);
+
+        Expression::Poison
     }
 }
